Rename Showcase → ExpandableTiles
The term "showcase" is only relevant on ListResources. It is labeled as
"featured analyses" on Splash. Renamed to be more self-descriptive.
