+++ static-site/src/components/ExpandableTiles/types.ts
